Confirm with positive test case. Little brittle but good enough

+++ src/Prenotes.Services/Things/Organization.cs
@@ -3,11 +3,17 @@ namespace Prenotes.Services.Things {
     
     public class Organization {
 
+        /// <summary>
+        /// Unique identifier
+        /// </summary>
         public readonly string name;
 
-        public readonly long created;
+        /// <summary>
+        /// When the orgization was registered
+        /// </summary>
+        public readonly int created;
 
-        public Organization(string name, long created) {
+        public Organization(string name, int created) {
             this.name = name;
             this.created = created;
         }
